backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "agents*", "routers*", "demo*"]

[project]
name = "et-investor-copilot-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.0",
    "yfinance>=0.2.40",
    "tavily-python>=0.3.0",
    "supabase>=2.4.0",
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]
